+++ core/__init__.py
@@ -7,7 +7,7 @@ __copyright__ = "Copyright (C) 2016-2018"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
-from sys import exit
+from sys import argv, exit
 from platform import system
 from .QtModules import QApplication
 from core.main_window import MainWindow
@@ -24,18 +24,28 @@ def main():
     """Startup function."""
     for info_str in INFO:
         print(info_str)
+
     print('-' * 7)
+
     if ARGUMENTS.test:
         print("All module loaded successfully.")
         exit(0)
-    app = QApplication([])
+
+    app = QApplication(argv)
+
+    # Force fusion style on Mac OS.
     if system() == 'Darwin':
         ARGUMENTS.fusion = True
+
+    # Fusion style.
     if ARGUMENTS.fusion:
         app.setStyle('fusion')
+
     splash = PyslvsSplash()
     splash.show()
+
     run = MainWindow()
     run.show()
     splash.finish(run)
+
     exit(app.exec())